# docs/swagger.yaml
basePath: /api/
definitions:
  apierr.ApiErr:
    properties:
      code:
        type: integer
      message: {}
    type: object
...
  dto.IdResponse:
    properties:
      id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      message:
        example: readme created successfully
        type: string
    required:
    - id
    type: object
  dto.LoginRequest:
    properties:
      login:
        maxLength: 80
        minLength: 1
        type: string
      password:
        minLength: 12
        type: string
    required:
    - login
    - password
    type: object
  dto.LoginResponse:
    properties:
      avatar:
        type: string
      id:
        type: string
      nickname:
        type: string
    required:
    - avatar
    - id
    - nickname
    type: object
...
  dto.RegisterRequest:
    properties:
      code:
        minLength: 6
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  dto.SendNewCodeRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: success
        type: string
    type: object
  dto.TemplateInfo:
    properties:
      description:
        minLength: 1
        type: string
      id:
        type: string
      image:
        type: string
      is_public:
        type: boolean
      last_update_time:
        type: string
      likes:
        minimum: 0
        type: integer
      num_of_users:
        minimum: 0
        type: integer
      title:
        minLength: 1
        type: string
    required:
    - description
    - id
    - image
    - is_public
    - last_update_time
    - likes
    - num_of_users
    - title
    type: object
...
  dto.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      nickname:
        minLength: 1
        type: string
      num_of_readmes:
        minimum: 0
        type: integer
      num_of_templates:
        minimum: 0
        type: integer
      templates:
        items:
          $ref: '#/definitions/dto.TemplateInfo'
        minItems: 0
        type: array
      time_of_register:
        type: string
    required:
    - avatar
    - email
    - id
    - nickname
    - num_of_readmes
    - num_of_templates
    - templates
    - time_of_register
    type: object
  dto.VerifyRequest:
    properties:
      email:
        type: string
      login:
        maxLength: 80
        minLength: 1
        type: string
      nickname:
        maxLength: 80
        minLength: 1
        type: string
      password:
        minLength: 12
        type: string
    required:
    - email
    - login
    - nickname
    - password
    type: object
...
host: localhost:3333
info:
  contact:
    email: readmeow@gmail.com
    name: READMEOW Support
  description: Documentation to READMEOW API
  title: REODMEOW API
  version: 1.0.0
paths:
  /api/auth/github:
    get:
      description: Start GitHub OAuth login flow. User will be redirected to Google
        for authentication, then back to your app. After successful login, the client
        will receive dto.LoginResponse.
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to GitHub OAuth
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Login via GitHub
      tags:
      - Auth
  /api/auth/google:
    get:
      description: Start Google OAuth login flow. User will be redirected to Google
        for authentication, then back to your app. After successful login, the client
        will receive dto.LoginResponse.
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to Google OAuth
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Login via Google
      tags:
      - Auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Log in a user with login and password
      parameters:
      - description: Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Login
      tags:
      - Auth
  /api/auth/logout:
    get:
      description: Logout a user
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Auth
  /api/auth/newcode:
    post:
      consumes:
      - application/json
      description: Sending a new code for user verification
      parameters:
      - description: SendNewCodeRequest Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendNewCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Send New Code
      tags:
      - Auth
  /api/auth/profile:
    get:
      description: Fetching the logged-in user's info
      produces:
      - application/json
      responses:
        "200":
          description: User info response
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      security:
      - ApiKeyAuth: []
      summary: Profile
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a verified user with email and verification code
      parameters:
      - description: Register Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.IdResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Register
      tags:
      - Auth
  /api/auth/verify:
    post:
      consumes:
      - application/json
      description: Verifying a user by sending a verification code
      parameters:
      - description: Verification Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/apierr.ApiErr'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierr.ApiErr'
      summary: Verify Email
      tags:
      - Auth
...
swagger: "2.0"